+++ sdk/python/v1alpha1/CHANGELOG.md
@@ -15,6 +15,7 @@ The versions coincide with releases on pip. Only major versions will be released
 
 ## [0.0.x](https://github.com/flux-framework/flux-operator/tree/main/sdk/python/v2alpha1) (0.0.x)
  - add status size variable (0.0.22)
+   - support maxSize to allow cluster scaling
  - support for staging (batchRaw) and batch submit (0.0.21)
  - tweak deletion logic to allow 404 response from get pods (0.0.2)
  - refactor of FluxOperator to include FluxMiniCluster to wrap it (0.0.19)